requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
sqlmodel
pydantic
sqlalchemy
pypdf
beautifulsoup4
lxml
alembic
typer

